+++ packages/sync-server/app/subCommands/loadSigner.js
@@ -8,7 +8,7 @@ import { Signer } from 'shared/models';
 import { loadCertificateIntoSigner } from '../integrations/Signer';
 import { initDatabase } from '../database';
 
-async function loadIcaoSigner({ signerCertificate }) {
+async function loadSigner({ signerCertificate }) {
   await initDatabase({ testMode: false });
   const signerFile = await fs.readFile(signerCertificate, 'utf8');
   const signerData = await loadCertificateIntoSigner(signerFile);
@@ -23,12 +23,12 @@ async function loadIcaoSigner({ signerCertificate }) {
   await pendingSigner.update(signerData);
   const start = moment(signerData.workingPeriodStart).format('YYYY-MM-DD');
   const end = moment(signerData.workingPeriodEnd).format('YYYY-MM-DD');
-  log.info(`Loaded ICAO Signer (${start} - ${end})`);
+  log.info(`Loaded Signer (${start} - ${end})`);
 
   process.exit(0);
 }
 
-export const loadIcaoSignerCommand = new Command('loadIcaoSigner')
+export const loadSignerCommand = new Command('loadSigner')
   .description('Loads an ICAO signer certificate into Tamanu')
   .requiredOption('-s, --signer-certificate <path>', 'Path to the signer certificate')
-  .action(loadIcaoSigner);
+  .action(loadSigner);